+++ src/main/java/org/bytedeco/javacv/FrameRecorder.java
@@ -112,6 +112,7 @@ public abstract class FrameRecorder implements Closeable {
     protected int sampleFormat, audioCodec, audioBitrate, sampleRate;
     protected double audioQuality = -1;
     protected boolean interleaved;
+    protected Charset charset = Charset.defaultCharset();
     protected Map<String, String> options = new HashMap<String, String>();
     protected Map<String, String> videoOptions = new HashMap<String, String>();
     protected Map<String, String> audioOptions = new HashMap<String, String>();
@@ -123,14 +124,6 @@ public abstract class FrameRecorder implements Closeable {
     protected int maxBFrames = -1;
     protected int trellis = -1;
     protected int maxDelay = -1;
-    protected Charset charset = Charset.defaultCharset();
-
-    public Charset getCharset() {
-        return charset;
-    }
-    public void setCharset(Charset charset) {
-        this.charset = charset;
-    }
 
     public String getFormat() {
         return format;
@@ -270,6 +263,13 @@ public abstract class FrameRecorder implements Closeable {
     }
     public void setInterleaved(boolean interleaved) {
         this.interleaved = interleaved;
+    }
+
+    public Charset getCharset() {
+        return charset;
+    }
+    public void setCharset(Charset charset) {
+        this.charset = charset;
     }
 
     public Map<String, String> getOptions() {