+++ src/main/java/org/bytedeco/javacv/FFmpegFrameGrabber.java
@@ -244,7 +244,6 @@ public class FFmpegFrameGrabber extends FrameGrabber {
             samples_convert_ctx = null;
         }
 
-        got_frame     = null;
         frameGrabbed  = false;
         frame         = null;
         timestamp     = 0;
@@ -382,8 +381,6 @@ public class FFmpegFrameGrabber extends FrameGrabber {
     private Buffer[]        samples_buf_out;
     private PointerPointer  plane_ptr, plane_ptr2;
     private AVPacket        pkt;
-    private int             sizeof_pkt;
-    private int[]           got_frame;
     private SwsContext      img_convert_ctx;
     private SwrContext      samples_convert_ctx;
     private int             samples_channels, samples_format, samples_rate;
@@ -554,7 +551,7 @@ public class FFmpegFrameGrabber extends FrameGrabber {
         AVDictionaryEntry entry = null;
         Map<String, String> metadata = new HashMap<String, String>();
         while ((entry = av_dict_get(oc.metadata(), "", entry, AV_DICT_IGNORE_SUFFIX)) != null) {
-            metadata.put(entry.key().getString(), entry.value().getString());
+            metadata.put(entry.key().getString(charset), entry.value().getString(charset));
         }
         return metadata;
     }
@@ -566,7 +563,7 @@ public class FFmpegFrameGrabber extends FrameGrabber {
         AVDictionaryEntry entry = null;
         Map<String, String> metadata = new HashMap<String, String>();
         while ((entry = av_dict_get(video_st.metadata(), "", entry, AV_DICT_IGNORE_SUFFIX)) != null) {
-            metadata.put(entry.key().getString(), entry.value().getString());
+            metadata.put(entry.key().getString(charset), entry.value().getString(charset));
         }
         return metadata;
     }
@@ -578,7 +575,7 @@ public class FFmpegFrameGrabber extends FrameGrabber {
         AVDictionaryEntry entry = null;
         Map<String, String> metadata = new HashMap<String, String>();
         while ((entry = av_dict_get(audio_st.metadata(), "", entry, AV_DICT_IGNORE_SUFFIX)) != null) {
-            metadata.put(entry.key().getString(), entry.value().getString());
+            metadata.put(entry.key().getString(charset), entry.value().getString(charset));
         }
         return metadata;
     }
@@ -588,7 +585,7 @@ public class FFmpegFrameGrabber extends FrameGrabber {
             return super.getMetadata(key);
         }
         AVDictionaryEntry entry = av_dict_get(oc.metadata(), key, null, 0);
-        return entry == null || entry.value() == null ? null : entry.value().getString();
+        return entry == null || entry.value() == null ? null : entry.value().getString(charset);
     }
 
     @Override public String getVideoMetadata(String key) {
@@ -596,7 +593,7 @@ public class FFmpegFrameGrabber extends FrameGrabber {
             return super.getVideoMetadata(key);
         }
         AVDictionaryEntry entry = av_dict_get(video_st.metadata(), key, null, 0);
-        return entry == null || entry.value() == null ? null : entry.value().getString();
+        return entry == null || entry.value() == null ? null : entry.value().getString(charset);
     }
 
     @Override public String getAudioMetadata(String key) {
@@ -604,7 +601,7 @@ public class FFmpegFrameGrabber extends FrameGrabber {
             return super.getAudioMetadata(key);
         }
         AVDictionaryEntry entry = av_dict_get(audio_st.metadata(), key, null, 0);
-        return entry == null || entry.value() == null ? null : entry.value().getString();
+        return entry == null || entry.value() == null ? null : entry.value().getString(charset);
     }
 
     /** default override of super.setFrameNumber implies setting
@@ -872,8 +869,6 @@ public class FFmpegFrameGrabber extends FrameGrabber {
         plane_ptr       = new PointerPointer(AVFrame.AV_NUM_DATA_POINTERS).retainReference();
         plane_ptr2      = new PointerPointer(AVFrame.AV_NUM_DATA_POINTERS).retainReference();
         pkt             = new AVPacket().retainReference();
-        sizeof_pkt      = pkt.sizeof();
-        got_frame       = new int[1];
         frameGrabbed    = false;
         frame           = new Frame();
         timestamp       = 0;
@@ -1154,6 +1149,9 @@ public class FFmpegFrameGrabber extends FrameGrabber {
                     initPictureRGB();
                 }
 
+                // Copy "metadata" fields
+                av_frame_copy_props(picture_rgb, picture);
+
                 // Convert the image into BGR or GRAY format that OpenCV uses
                 img_convert_ctx = sws_getCachedContext(img_convert_ctx,
                         video_c.width(), video_c.height(), video_c.pix_fmt(),
@@ -1389,7 +1387,6 @@ public class FFmpegFrameGrabber extends FrameGrabber {
                 }
 
                 // Did we get a video frame?
-                got_frame[0] = 0;
                 while (!done) {
                     ret = avcodec_receive_frame(video_c, picture);
                     if (ret == AVERROR_EAGAIN() || ret == AVERROR_EOF()) {
@@ -1400,9 +1397,11 @@ public class FFmpegFrameGrabber extends FrameGrabber {
                             break;
                         }
                     } else if (ret < 0) {
-                        throw new Exception("avcodec_receive_frame() error " + ret + ": Error during video decoding.");
+                        // Ignore errors to emulate the behavior of the old API
+                        // throw new Exception("avcodec_receive_frame() error " + ret + ": Error during video decoding.");
+                        readPacket = true;
+                        break;
                     }
-                    got_frame[0] = 1;
 
                     if (!keyFrames || picture.pict_type() == AV_PICTURE_TYPE_I) {
                         long pts = picture.best_effort_timestamp();
@@ -1419,7 +1418,8 @@ public class FFmpegFrameGrabber extends FrameGrabber {
                         done = true;
                         frame.timestamp = timestamp;
                         frame.keyFrame = picture.key_frame() != 0;
-                        frame.type =Type.VIDEO;
+                        frame.pictType = (char)av_get_picture_type_char(picture.pict_type());
+                        frame.type = Type.VIDEO;
                     }
                 }
             } else if (doAudio && audio_st != null && pkt_stream_ind == audio_st.index()) {
@@ -1427,21 +1427,23 @@ public class FFmpegFrameGrabber extends FrameGrabber {
                 if (readPacket) {
                     ret = avcodec_send_packet(audio_c, pkt);
                     if (ret < 0) {
-                        throw new Exception("avcodec_send_packet() error " + ret + ": Error sending an audio packet for decoding.");
+                        // Ignore errors to emulate the behavior of the old API
+                        // throw new Exception("avcodec_send_packet() error " + ret + ": Error sending an audio packet for decoding.");
                     }
                 }
 
                 // Did we get an audio frame?
-                got_frame[0] = 0;
                 while (!done) {
                     ret = avcodec_receive_frame(audio_c, samples_frame);
                     if (ret == AVERROR_EAGAIN() || ret == AVERROR_EOF()) {
                         readPacket = true;
                         break;
                     } else if (ret < 0) {
-                        throw new Exception("avcodec_receive_frame() error " + ret + ": Error during audio decoding.");
+                        // Ignore errors to emulate the behavior of the old API
+                        // throw new Exception("avcodec_receive_frame() error " + ret + ": Error during audio decoding.");
+                        readPacket = true;
+                        break;
                     }
-                    got_frame[0] = 1;
 
                     long pts = samples_frame.best_effort_timestamp();
                     AVRational time_base = audio_st.time_base();